+++ TypographyKit/Classes/TypographyColor.swift
@@ -15,12 +15,18 @@ public indirect enum TypographyColor {
     
     // MARK: - Type definitions
     private struct RegEx {
-        static let hexColor1 = "#[a-zA-Z0-9]{6}"
-        static let hexColor2 = "[a-zA-Z0-9]{6}"
+        static let hexWithoutAlphaColor = "#[a-zA-Z0-9]{6}"
+        static let hexWithAlphaColor = "#[a-zA-Z0-9]{8}"
+        static let hexWAlphaAndWoHashColor = "[a-zA-Z0-9]{8}"
+        static let hexWoAlphaAndWoHashColor = "[a-zA-Z0-9]{6}"
         // swiftlint:disable:next line_length
-        static let rgbColor1 = "rgb\\(([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5])\\)"
+        static let rgbColorWithPrefix = "rgb\\(([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5])\\)"
         // swiftlint:disable:next line_length
-        static let rgbColor2 = "\\(([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5])\\)"
+        static let rgbaColorWithPrefix = "rgba\\(([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5])\\)"
+        // swiftlint:disable:next line_length
+        static let rgbColorWithoutPrefix = "\\(([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5])\\)"
+        // swiftlint:disable:next line_length
+        static let rgbaColorWithoutPrefix = "\\(([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]),([01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5])\\)"
     }
     
     // MARK: - Cases
@@ -39,7 +45,8 @@ public indirect enum TypographyColor {
     case purple
     case brown
     case clear
-    case hex(string: String)
+    case hexWithoutAlpha(string: String)
+    case hexWithAlpha(string: String)
     case named(string: String)
     case rgb(red: Float, green: Float, blue: Float)
     case rgba(red: Float, green: Float, blue: Float, alpha: Float)
@@ -48,21 +55,23 @@ public indirect enum TypographyColor {
     
     // MARK: - Properties
     static var colorNameMap: [String: UIColor] {
-        return ["black": .black,
-                "darkGray": .darkGray,
-                "lightGray": .lightGray,
-                "white": .white,
-                "gray": .gray,
-                "red": .red,
-                "green": .green,
-                "blue": .blue,
-                "cyan": .cyan,
-                "yellow": .yellow,
-                "magenta": .magenta,
-                "orange": .orange,
-                "purple": .purple,
-                "brown": .brown,
-                "clear": .clear]
+        return [
+            "black": .black,
+            "darkGray": .darkGray,
+            "lightGray": .lightGray,
+            "white": .white,
+            "gray": .gray,
+            "red": .red,
+            "green": .green,
+            "blue": .blue,
+            "cyan": .cyan,
+            "yellow": .yellow,
+            "magenta": .magenta,
+            "orange": .orange,
+            "purple": .purple,
+            "brown": .brown,
+            "clear": .clear
+        ]
     }
     
     public var cgColor: CGColor {
@@ -119,9 +128,12 @@ public indirect enum TypographyColor {
             if let color = UIColor(named: colorName) {
                 return color
             }
-            return TypographyColor.colorNameMap[colorName]! // Previously validated
-        case .hex(let hexString):
-            return TypographyColor.parseHex(hexString: hexString)!.uiColor // Previously validated
+            // Force unwrap works because has been validated. TODO: Replace force unwrap.
+            return TypographyColor.colorNameMap[colorName]!
+        case let .hexWithoutAlpha(hexString),
+            let .hexWithAlpha(hexString):
+            // Force unwrap works because has been validated. TODO: Replace force unwrap.
+            return TypographyColor.parseHex(hexString: hexString)!.uiColor
         case .rgb(let red, let green, let blue):
             return UIColor(red: CGFloat(red), green: CGFloat(green), blue: CGFloat(blue), alpha: 1.0)
         case .rgba(let red, let green, let blue, let alpha):
@@ -130,7 +142,6 @@ public indirect enum TypographyColor {
             guard #available(iOS 13, *) else {
                 return colorDictionary[.light]?.uiColor ?? .black
             }
-            
             return UIColor { (traitCollection) -> UIColor in
                 let key = TypographyInterfaceStyle(style: traitCollection.userInterfaceStyle)
                 let color = colorDictionary[key] ?? colorDictionary[.light]
@@ -150,18 +161,33 @@ public indirect enum TypographyColor {
         }
         
         switch string {
-        case RegEx.hexColor1, RegEx.hexColor2:
+        case RegEx.hexWithoutAlphaColor, RegEx.hexWoAlphaAndWoHashColor:
             guard TypographyColor.parseHex(hexString: string) != nil else { // Check can be converted to a color
                 return nil
             }
-            self = .hex(string: string)
-        case RegEx.rgbColor1, RegEx.rgbColor2:
+            self = .hexWithoutAlpha(string: string)
+        case RegEx.hexWithAlphaColor, RegEx.hexWAlphaAndWoHashColor:
+            guard TypographyColor.parseHex(hexString: string) != nil else { // Check can be converted to a color
+                return nil
+            }
+            self = .hexWithAlpha(string: string)
+        case RegEx.rgbColorWithPrefix, RegEx.rgbColorWithoutPrefix:
             let rgbValues = type(of: self).rgbValues(from: string)
             guard rgbValues.count == 3, let red = Float(rgbValues[0]), let green = Float(rgbValues[1]),
-                let blue = Float(rgbValues[2]) else {
-                    return nil
+                  let blue = Float(rgbValues[2]) else {
+                return nil
             }
             self = .rgb(red: red / 255.0, green: green / 255.0, blue: blue / 255.0)
+        case RegEx.rgbaColorWithPrefix, RegEx.rgbaColorWithoutPrefix:
+            let rgbValues = type(of: self).rgbValues(from: string)
+            guard rgbValues.count == 4,
+                  let red = Float(rgbValues[0]),
+                  let green = Float(rgbValues[1]),
+                  let blue = Float(rgbValues[2]),
+                  let alpha = Float(rgbValues[3]) else {
+                return nil
+            }
+            self = .rgba(red: red / 255.0, green: green / 255.0, blue: blue / 255.0, alpha: alpha / 255.0)
         default:
             return nil
         }
@@ -174,24 +200,31 @@ private extension TypographyColor {
     
     /// Parses a hexadecimal string to a color if possible
     private static func parseHex(hexString: String) -> TypographyColor? {
-        let unparsed = hexString.hasPrefix("#")
-            ? String(hexString[hexString.index(after: hexString.startIndex)...])
-            : hexString
+        // Obtain the hex value without the # prefix.
+        let hexValue = hexString.hasPrefix("#")
+        ? String(hexString[hexString.index(after: hexString.startIndex)...])
+        : hexString
         
-        let redComponentIdx = unparsed.startIndex,
-        greenComponentIdx = unparsed.index(unparsed.startIndex, offsetBy: 2),
-        blueComponentIdx = unparsed.index(unparsed.startIndex, offsetBy: 4)
-        
-        let redComponent = unparsed[redComponentIdx..<greenComponentIdx],
-        greenComponent = unparsed[greenComponentIdx..<blueComponentIdx],
-        blueComponent = unparsed[blueComponentIdx..<unparsed.endIndex]
-        
-        if let rValue = UInt(redComponent, radix: 16), let gValue = UInt(greenComponent, radix: 16),
-            let bValue = UInt(blueComponent, radix: 16) {
+        let redComponentIdx = hexValue.startIndex,
+            greenComponentIdx = hexValue.index(hexValue.startIndex, offsetBy: 2),
+            blueComponentIdx = hexValue.index(hexValue.startIndex, offsetBy: 4),
+            alphaIdx = hexValue.index(hexValue.startIndex, offsetBy: 6)
+        let redComponent = hexValue[redComponentIdx..<greenComponentIdx],
+            greenComponent = hexValue[greenComponentIdx..<blueComponentIdx],
+            blueComponent = hexValue[blueComponentIdx..<alphaIdx]
+        let alphaComponent = hexValue[alphaIdx..<hexValue.endIndex]
+        if let rValue = UInt(redComponent, radix: 16),
+           let gValue = UInt(greenComponent, radix: 16),
+           let bValue = UInt(blueComponent, radix: 16) {
             let red = Float(rValue) / 255.0
             let green = Float(gValue) / 255.0
             let blue = Float(bValue) / 255.0
-            return .rgb(red: red, green: green, blue: blue)
+            if alphaComponent.isEmpty {
+                return .rgb(red: red, green: green, blue: blue)
+            } else if let alphaBase10Val = UInt(alphaComponent, radix: 16) {
+                let rgba32AlphaVal = Float(alphaBase10Val) / 255.0
+                return .rgba(red: red, green: green, blue: blue, alpha: rgba32AlphaVal)
+            }
         }
         return nil
     }
@@ -200,11 +233,15 @@ private extension TypographyColor {
         var colorValues: [String] = []
         do {
             let colorComponentPattern = "[01]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5]"
-            let colorComponentRegEx = try NSRegularExpression(pattern: colorComponentPattern,
-                                                              options: .dotMatchesLineSeparators)
-            let matches = colorComponentRegEx.matches(in: string,
-                                                      options: [],
-                                                      range: NSRange(location: 0, length: string.count))
+            let colorComponentRegEx = try NSRegularExpression(
+                pattern: colorComponentPattern,
+                options: .dotMatchesLineSeparators
+            )
+            let matches = colorComponentRegEx.matches(
+                in: string,
+                options: [],
+                range: NSRange(location: 0, length: string.count)
+            )
             for match in matches {
                 let matchEndIndex = match.range.location + match.range.length
                 let startIdx = string.index(string.startIndex, offsetBy: match.range.location)