+++ TypographyKit/Classes/UIColorAdditions.swift
@@ -18,12 +18,32 @@ extension UIColor {
         var alpha: CGFloat = 0.0
     }
     
-    public enum Lightness {
-        private static let lightScalingFactor: Double       = 1.25
-        private static let lighterScalingFactor: Double     = 1.5
-        private static let lightestScalingFactor: Double    = 1.75
-        private static let whiteScalingFactor: Double       = Double.greatestFiniteMagnitude
+    public enum Shade: CaseIterable {
+        public static var allCases: [UIColor.Shade] = [
+            .black,
+            .dark,
+            .darker,
+            .darkest,
+            .light,
+            .lighter,
+            .lightest,
+            .white
+        ]
+        
+        private static let blackScalingFactor: Double    = 0.0
+        private static let darkScalingFactor: Double     = 0.75
+        private static let darkerScalingFactor: Double   = 0.5
+        private static let darkestScalingFactor: Double  = 0.25
+        private static let lightScalingFactor: Double    = 1.25
+        private static let lighterScalingFactor: Double  = 1.5
+        private static let lightestScalingFactor: Double = 1.75
+        private static let whiteScalingFactor: Double    = Double.greatestFiniteMagnitude
 
+        case black
+        case darkness(Double)
+        case dark
+        case darker
+        case darkest
         case lightness(scalingFactor: Double)
         case light
         case lighter
@@ -32,44 +52,24 @@ extension UIColor {
 
         fileprivate var scale: Double {
             switch self {
-            case .lightness(let scalingFactor):
+            case let .darkness(scalingFactor), let .lightness(scalingFactor):
                 return scalingFactor
-            case .light:
-                return Lightness.lightScalingFactor
-            case .lighter:
-                return Lightness.lighterScalingFactor
-            case .lightest:
-                return Lightness.lightestScalingFactor
-            case .white:
-                return Lightness.whiteScalingFactor
-            }
-        }
-    }
-
-    public enum Darkness {
-        private static let darkScalingFactor: Double    = 0.75
-        private static let darkerScalingFactor: Double  = 0.5
-        private static let darkestScalingFactor: Double = 0.25
-        private static let blackScalingFactor: Double   = 0.0
-
-        case darkness(Double)
-        case dark
-        case darker
-        case darkest
-        case black
-
-        fileprivate var scale: Double {
-            switch self {
-            case .darkness(let scalingFactor):
-                return scalingFactor
-            case .dark:
-                return Darkness.darkScalingFactor
-            case .darker:
-                return Darkness.darkerScalingFactor
-            case .darkest:
-                return Darkness.darkestScalingFactor
             case .black:
-                return Darkness.blackScalingFactor
+                return Shade.blackScalingFactor
+            case .dark:
+                return Shade.darkScalingFactor
+            case .darker:
+                return Shade.darkerScalingFactor
+            case .darkest:
+                return Shade.darkestScalingFactor
+            case .light:
+                return Shade.lightScalingFactor
+            case .lighter:
+                return Shade.lighterScalingFactor
+            case .lightest:
+                return Shade.lightestScalingFactor
+            case .white:
+                return Shade.whiteScalingFactor
             }
         }
     }
@@ -77,39 +77,38 @@ extension UIColor {
     private func colorComponents(scalingFactor: CGFloat) -> ColorComponents {
         var red: CGFloat = 0.0, green: CGFloat = 0.0, blue: CGFloat = 0.0, alpha: CGFloat = 0.0
         self.getRed(&red, green: &green, blue: &blue, alpha: &alpha)
-        return ColorComponents(red: red * scalingFactor, green: green * scalingFactor, blue: blue * scalingFactor,
-                               alpha: alpha)
-    }
-    
-    private func darker(darkness: Double) -> UIColor {
-        guard darkness <= 1.0 else { return self }
-        var components = colorComponents(scalingFactor: CGFloat(darkness))
-        components.red = CGFloat.maximum(components.red, 0.0)
-        components.green = CGFloat.maximum(components.green, 0.0)
-        components.blue = CGFloat.maximum(components.blue, 0.0)
-        return UIColor(red: components.red, green: components.green, blue: components.blue, alpha: components.alpha)
-    }
-    
-    private func lighter(lightness: Double) -> UIColor {
-        guard lightness >= 1.0 else { return self }
-        var components = colorComponents(scalingFactor: CGFloat(lightness))
-        components.red = CGFloat.minimum(components.red, 1.0)
-        components.green = CGFloat.minimum(components.green, 1.0)
-        components.blue = CGFloat.minimum(components.blue, 1.0)
-        return UIColor(red: components.red, green: components.green, blue: components.blue, alpha: components.alpha)
+        return ColorComponents(
+            red: red * scalingFactor,
+            green: green * scalingFactor,
+            blue: blue * scalingFactor,
+            alpha: alpha
+        )
     }
 
-    public func shade(_ lightness: Lightness) -> UIColor {
-        switch lightness {
-        case .white:
-            return UIColor.white
-        default:
-            return self.lighter(lightness: lightness.scale)
+    public func shade(_ shade: Shade) -> UIColor {
+        var components = colorComponents(scalingFactor: CGFloat(shade.scale))
+        switch shade {
+        case .black, .dark, .darker, .darkest, .darkness:
+            if shade.scale > 1.0 {
+                return self
+            }
+            components.red = CGFloat.maximum(components.red, 0.0)
+            components.green = CGFloat.maximum(components.green, 0.0)
+            components.blue = CGFloat.maximum(components.blue, 0.0)
+        case .white, .light, .lighter, .lightest, .lightness:
+            if shade.scale < 1.0 {
+                return self
+            }
+            components.red = CGFloat.minimum(components.red, 1.0)
+            components.green = CGFloat.minimum(components.green, 1.0)
+            components.blue = CGFloat.minimum(components.blue, 1.0)
         }
-    }
-
-    public func shade(_ darkness: Darkness) -> UIColor {
-        return self.darker(darkness: darkness.scale)
+        return UIColor(
+            red: components.red,
+            green: components.green,
+            blue: components.blue,
+            alpha: components.alpha
+        )
     }
 
     public func shade(_ name: String) -> UIColor {